[ICRDMA] Move cq actor into separate library. EXT-1618 (#27501)
[ICRDMA] Move cq actor into separate library.
Add dummy implementation for unsupported platforms. EXT-1618
