+++ ydb/library/actors/interconnect/rdma/ut/ya.make
@@ -21,6 +21,7 @@ PEERDIR(
     ydb/library/actors/core
     ydb/library/actors/interconnect
     ydb/library/actors/interconnect/rdma
+    ydb/library/actors/interconnect/rdma/cq_actor
     ydb/library/actors/testlib
 )
